refactor: divide module
