+++ pokemon_ai/dqn/value_function_agent.py
@@ -1,6 +1,3 @@
-import logging
-from collections import deque
-from dataclasses import dataclass
 from random import random, sample
 from typing import Optional
 
@@ -9,27 +6,11 @@ from sklearn.neural_network import MLPRegressor
 
 import pokemon_ai.simulator.moves as m
 import pokemon_ai.simulator.pokedex as p
+from pokemon_ai.dqn.experience import Experience
 from pokemon_ai.simulator.battle_simplified import Battle
 from pokemon_ai.simulator.player import Action, Player
 
-# from sklearn.pipeline import Pipeline
-# from sklearn.preprocessing import StandardScaler
-
-
 GAMMA = 0.9
-
-
-# TODO: use dataclass
-# Experience = namedtuple("Experience", ["state", "action", "reward", "next_state", "done"])
-
-
-@dataclass
-class Experience:
-    state: list[int]
-    action: Action
-    reward: float
-    next_state: list[int]
-    done: bool
 
 
 class StupidRandomPlayer(Player):
@@ -131,76 +112,3 @@ def build_random_team() -> list[p.Pokemon]:
         p.Starmie([m.Surf()]),
     ]
     return sample(pokemons, 2)
-
-
-class Trainer:
-    step = 0
-    experiences = deque(maxlen=1024)
-
-    def __init__(self, episodes: int, model: Optional[MLPRegressor] = None, epsilon: float = 0.2):
-        self.agent = ValueFunctionAgent(model=model, epsilon=epsilon)
-        self.episodes = episodes
-
-    def train(self):
-        win_count_total = 0
-        win_count_of_100 = 0
-        for episode in range(0, self.episodes):
-            self.agent.reset()
-            battle = Battle(self.agent.learner, self.agent.opponent)
-            battle.validate()
-            logging.info(battle)
-            while True:
-                logging.info(f"=== turn {battle.turn} ===")
-                current_state = battle.to_array()
-
-                current_learner_pokemon_count = len(battle.player1.get_available_pokemons())
-                current_opponent_pokemon_count = len(battle.player2.get_available_pokemons())
-
-                action, _ = battle.forward_step()
-                winner = battle.get_winner()
-
-                next_learner_pokemon_count = len(battle.player1.get_available_pokemons())
-                next_opponent_pokemon_count = len(battle.player2.get_available_pokemons())
-
-                if action is not None:
-                    reward = 0
-                    if winner == self.agent.learner:
-                        logging.info(f"learner won the battle!")
-                        reward = 1
-                    if winner == self.agent.opponent:
-                        logging.info(f"learner lost...")
-                        reward = -1
-                    if battle.turn > 500:
-                        logging.info("battle is too long")
-                        reward = -0.1
-                    if current_opponent_pokemon_count > next_opponent_pokemon_count:
-                        reward += 0.3
-                    if current_learner_pokemon_count > next_learner_pokemon_count:
-                        reward -= 0.3
-                    self.experiences.append(
-                        Experience(
-                            state=current_state,
-                            action=action,
-                            reward=reward,
-                            next_state=battle.to_array(),
-                            done=winner is not None,
-                        )
-                    )
-                self.step += 1
-                logging.info(battle)
-                if winner is not None or battle.turn > 500:
-                    if winner == self.agent.learner:
-                        win_count_total += 1
-                        win_count_of_100 += 1
-                    break
-            if len(self.experiences) > 64:
-                self.agent.update(sample(self.experiences, 64))
-            if episode > 0 and episode % 100 == 0:
-                print(f"=============")
-                print(f"episode {episode}")
-                print(f"Win Rate of 100: {win_count_of_100 / 100}")
-                print(f"Total Win Rate: {win_count_total / episode}")
-                win_count_of_100 = 0
-        print(f"=======================================")
-        print(f"Total Win Count: {win_count_total}")
-        print(f"Total Win Rate: {win_count_total / self.episodes}")
